food_recommendation_agency/requirements.txt
agency-swarm
python-dotenv>=0.19.0
pydantic
openai>=1.0.0 